+++ integration_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/csmith/containuum"
+	"github.com/csmith/containuum/v2"
 	"github.com/sebdah/goldie/v2"
 	"github.com/stretchr/testify/require"
 )